[addons][gui] move addon gui source files to own folder
This is done to have more clean and all GUI only related parts there
in own folder.

Also comes with another request a new helper class about addon GUI to handle
same parts on several places.
