+++ xbmc/addons/ContextMenus.cpp
@@ -9,7 +9,6 @@
 #include "ContextMenus.h"
 
 #include "AddonManager.h"
-#include "GUIDialogAddonInfo.h"
 #include "Repository.h"
 #include "RepositoryUpdater.h"
 #include "ServiceBroker.h"
