+++ xbmc/addons/gui/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(SOURCES GUIDialogAddonInfo.cpp
+            GUIViewStateAddonBrowser.cpp
+            GUIWindowAddonBrowser.cpp)
+
+set(HEADERS GUIDialogAddonInfo.h
+            GUIViewStateAddonBrowser.h
+            GUIWindowAddonBrowser.h)
+
+core_add_library(addons_gui)
