+++ xbmc/addons/gui/GUIDialogAddonInfo.h
