+++ xbmc/addons/gui/GUIViewStateAddonBrowser.cpp
