+++ xbmc/addons/gui/GUIViewStateAddonBrowser.h
