+++ xbmc/games/controllers/windows/GUIControllerWindow.cpp
@@ -14,8 +14,8 @@
 #include "GUIFeatureList.h"
 #include "ServiceBroker.h"
 #include "addons/AddonManager.h"
-#include "addons/GUIWindowAddonBrowser.h"
 #include "addons/IAddon.h"
+#include "addons/gui/GUIWindowAddonBrowser.h"
 #include "cores/RetroPlayer/guibridge/GUIGameRenderManager.h"
 #include "cores/RetroPlayer/guibridge/GUIGameSettingsHandle.h"
 #include "games/addons/GameClient.h"
