+++ xbmc/programs/GUIWindowPrograms.cpp
@@ -13,7 +13,7 @@
 #include "GUIPassword.h"
 #include "ServiceBroker.h"
 #include "Util.h"
-#include "addons/GUIDialogAddonInfo.h"
+#include "addons/gui/GUIDialogAddonInfo.h"
 #include "dialogs/GUIDialogMediaSource.h"
 #include "guilib/GUIComponent.h"
 #include "guilib/GUIWindowManager.h"
