+++ xbmc/settings/dialogs/GUIDialogContentSettings.cpp
@@ -10,7 +10,7 @@
 
 #include "ServiceBroker.h"
 #include "addons/AddonSystemSettings.h"
-#include "addons/GUIWindowAddonBrowser.h"
+#include "addons/gui/GUIWindowAddonBrowser.h"
 #include "addons/settings/GUIDialogAddonSettings.h"
 #include "dialogs/GUIDialogKaiToast.h"
 #include "dialogs/GUIDialogSelect.h"
