+++ xbmc/view/GUIViewState.cpp
@@ -17,8 +17,8 @@
 #include "ViewDatabase.h"
 #include "addons/Addon.h"
 #include "addons/AddonManager.h"
-#include "addons/GUIViewStateAddonBrowser.h"
 #include "addons/PluginSource.h"
+#include "addons/gui/GUIViewStateAddonBrowser.h"
 #include "dialogs/GUIDialogSelect.h"
 #include "events/windows/GUIViewStateEventLog.h"
 #include "filesystem/AddonsDirectory.h"
